refactor: 지도상태 관찰 / 변경에 따른 지도 상태 커스텀훅 구분 (#173)
